+++ lib/metro/stubs/interceptor_stub.dart
@@ -1,0 +1,23 @@
+import 'package:recase/recase.dart';
+
+/// This stub is used to create an Interceptor.
+String interceptorStub({required ReCase interceptorName}) => '''
+import 'package:nylo_framework/nylo_framework.dart';
+
+class ${interceptorName.pascalCase}Interceptor extends Interceptor {
+  @override
+  void onRequest(RequestOptions options, RequestInterceptorHandler handler) {
+    return super.onRequest(options, handler);
+  }
+
+  @override
+  void onResponse(Response response, ResponseInterceptorHandler handler) {
+    handler.next(response);
+  }
+
+  @override
+  void onError(DioException dioException, ErrorInterceptorHandler handler) {
+    handler.next(dioException);
+  }
+}
+''';