+++ lib/metro/stubs/page_stub.dart
@@ -1,25 +1,29 @@
 import 'package:recase/recase.dart';
 
 /// This stub is used to create a new Page.
-String pageStub({required ReCase pageName}) => '''
+String pageStub({required ReCase className}) => '''
 import 'package:flutter/material.dart';
 import 'package:nylo_framework/nylo_framework.dart';
 
-class ${pageName.pascalCase}Page extends NyPage {
+class ${className.pascalCase}Page extends NyStatefulWidget {
+  static const path = '/${className.paramCase}';
+  
+  ${className.pascalCase}Page() : super(path, child: _${className.pascalCase}PageState());
+}
 
-  static String path = '/${pageName.paramCase}';
+class _${className.pascalCase}PageState extends NyState<${className.pascalCase}Page> {
 
   @override
   init() async {
-    // To update the state, 
-    // use: refreshPage(setState: () { }); 
+    super.init();
+
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text("${pageName.titleCase}")
+        title: Text("${className.titleCase}")
       ),
       body: SafeArea(
          child: Container(),